Moved UiDispatcher into the Rubberduck.Parsing assembly to make it accessible from there.
