+++ RetailCoder.VBE/UI/Command/FindAllImplementationsCommand.cs
@@ -7,6 +7,7 @@ using NLog;
 using Rubberduck.Common;
 using Rubberduck.Parsing.Grammar;
 using Rubberduck.Parsing.Symbols;
+using Rubberduck.Parsing.UIContext;
 using Rubberduck.Parsing.VBA;
 using Rubberduck.UI.Command.MenuItems;
 using Rubberduck.UI.Controls;
