+++ RetailCoder.VBE/UI/Command/ShowParserErrorsCommand.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 using System.Windows.Input;
 using NLog;
 using Rubberduck.Parsing.Symbols;
+using Rubberduck.Parsing.UIContext;
 using Rubberduck.Parsing.VBA;
 using Rubberduck.UI.Command.MenuItems;
 using Rubberduck.UI.Controls;
